Checkpoint commit.
At this moment, the silly idea of supporting a few Unix commands has been
tossed in favor of a file listing all commands deemed worthy of pass-through
to the user's shell.  Seems to be working okay, and the implementation of
copy has been seperated into two phases.  Phase one is now in progress...


git-svn-id: https://svn.code.sf.net/p/distribulator/code/trunk@37 9669c016-0312-0410-afc9-92ac2c5c32b3